Add bootstrap code to take care of the boot sequence and prepare the cpu before jumping to rust code.
This code will setup a GDT, enable 32-bit Protected Mode, setup segment
registers and a small stack in low memory.

For now the cpu operates in 32-bit mode. 64-bit will be implemented
later.

The jump to the rust code is not yet implemented. More work on the rust
build is necessary before being able to jump to it.