Boostrap: Add new 32-bit routine allowing one to call BIOS function from 32-bit Protected Mode.
This routine takes as argument a BIOS Call Packet (BCP) which indicate:
        - The interrupt to execute in real-mode. E.g. The BIOS function.
        - The value of all the GP registers (including 32-bit regs).
This function will take care of switching back to real-mode, set the
registers to the requested values, execute the requested BIOS function,
write the output (e.g. values of registers) back into the BCP (including
the value of the Carry Flag) and return back to 32-bit Protected Mode to
the caller.