Bootstrap: Add a physical frame allocator.
The main data structure used by the allocator is a linked list where
each node describes a block of contiguous free frames. Each node has the
following layout:
 - (QWORD) Address of the next entry/node in the linked list.
 - (QWORD) Start address of the block. That is the address of the first
   frame of the block.
 - (QWORD) Size of the block in number of frames.
A global variable (FRAME_ALLOC_HEAD) maintains a pointer to the head of
the linked-list.
Allocating a frame is straight-forward, one simply needs to look at the
first entry of the linked list, the allocated frame is the first frame
of the block described by the entry. The allocating procedure then
update the start address and size fields of the entry. If this entry is
empty then it is removed from the list.
Free-ing frames is not yet implemented as there are no use case for that
yet.