Bootstrap: Add initialization and configuration routines for initializing the VESA video mode.
The new routine init_video will use the BIOS VESA functions to find a
video mode corresponding to the requested resolution and depth (.sets in
the video.S) and, once found, set it as the current video mode. If no
such mode exists then it panics.

This routine is not yet called by the bootstrap because the bootstrap
still relies on the VGA text buffer which becomes unavailable after
changing the video mode. The call to init_video will be added in a later
commit once bootstrap logging has been moved to serial output instead.