Boostrap/Interrupts: Add routines to initialize interrupt handling.
These new routine in the interrupt.S file will dynamically allocate an
IDT, and interrupt handlers.

All interrupt handlers call a generic interrupt handler which compute
the interrupt vector, save the registers, build an interrupt frame and
start the processing of the interrupt. Once the interrupt is processed
the generic handler will restore the registers and return to the
interrupted context.