Bootstrap/Interrupts: Add a callback system for interrupts.
Instead of processing any vector, the generic_handler delegates
processing to `interrupt callbacks`. An interrupt callback is a routine
taking as argument  a pointer to an interrupt frame.
Interrupt callbacks are stored in a table indexed by vector number. When
an interrupt is raised, the generic_handler looks up the table and call
the callback, if any, passing as argument (RDI) a pointer to the
interrupt frame.