Bootstrap: Add TSC frequency computation.
In this project, we use the TSC as a time-keeping counter, this means
that we need the TSC to have a _constant_ frequency no matter the state
of the CPU (running, idle, freq scaling, ...). Fortunately all recent
Intel processors have this feature, this is called "Invariant TSC" and
support can be asserted using CPUID.
Unfortunately getting the TSC's frequency is an adventure on itself.
Intel provides multiple ways to achieve this, but none seem to be
portable and in reality it depends on the processor's model.
Additionally Qemu gets in our way by hiding some CPUID information and
MSRs from the guest (even though we are using the host cpu model, ...).
If all fails we can use the PIT to calibrate the TSC's frequency. This
is the least precise method but it yields decent results, and it works
in Qemu VMs.

Here are the three methods to get the TSC's frequency:
 1. Using CPUID.15H. This execution to CPUID will give enough
    information to compute the frequency. Unfortunately, not all cpus
    implement CPUID.15H. The precision using this method is high.
 2. Using the MSR_PLATFORM_INFO (0xCE). This MSR gives the "Maximum
    Non-Turbo Ratio". This ration can be used to compute the TSC's
    frequency. Using this method precision is high.
 3. Using the PIT. Precision is medium to poor but is usually
    sufficient.

Upon calling init_tsc, each method will be tried (in the order above) to
find the TSC's frequency.