Bootstrap/Testing: Add testing framework for 64-bit mode tests.
We reuse the test table. Each entry now has a byte indicating the mode
it should be ran in. If the value is 32 then this is a 32-bit test, if
the value is 64 then this is a 64-bit test.
Testing is divided into two routines: run_tests32 (previously run_tests)
runs 32-bit tests while run_tests64 runs 64-bit ones.