Bootstrap: Add support for syscalls.
Syscalls in Krayte are done through software interrupts for now. In the
future we might add support for the SYSCALL/SYSENTER instructions, but
for now we try to keep things simple (interrupts are well tested) and
portable (at least for syscalls).

Syscalls through software interrupts use the following conventions:
 - Vector is INTERRUPT_SYSCALL_VEC.
 - RAX contains the syscall number.
 - Parameters of the syscall are passed through registers in the
   following order:
     1. RDI
     2. RSI
     3. RDX
     4. RCX
     5. R8
     6. R9
 - For now a syscall can only take up to 6 parameters. Passing extra
   parameters through the stack can be added later. There is no use
   cases for more than 6 parameters for now.
 - RAX will contains the return value of the syscall.

Syscall numbers are defined in consts.h