Change CMakeLists.txt to produce different executables depending on the build type.
For DEBUG builds, CMake will produce an executable that can be run on
the development machine. In this mode the STD library is allowed.
For RELEASE builds, CMake will produce an executable meant to be run
directly on baremetal. STD library is disabled.

Both types will produce the same executable: Krayte to avoid confusion.